move paint property setup to separate function/file
