+++ public/src/app.js
@@ -1,5 +1,6 @@
 import { state } from "./initialState.js";
 import { dimensions } from "./constants.js";
+import { setupPaintProperties } from "./utils/setupPaintProperties.js";
 
 const app = (s) => {
   const socket = io.connect("http://localhost:3000");
@@ -103,37 +104,7 @@ const app = (s) => {
   };
 
   s.mouseDragged = () => {
-    state.rainbowCounter += 1 * state.gui.rainbowSpeed;
-
-    const hue = Math.floor(state.rainbowCounter % 360);
-    const saturation = s.map(s.mouseY, 0, s.windowHeight, 100, 75);
-    const brightness = s.map(s.mouseY, 0, s.windowHeight, 100, 50);
-
-    const rainbowFill = s.color(`hsb(${hue}, ${saturation}%, ${brightness}%)`);
-    rainbowFill.setAlpha(state.gui.fillOpacity);
-
-    const rainbowStroke = s.color(
-      `hsb(${hue}, ${saturation}%, ${brightness}%)`
-    );
-    rainbowStroke.setAlpha(state.gui.strokeOpacity);
-
-    const paintProperties = {
-      x: s.mouseX,
-      y: s.mouseY,
-      fillColor: state.gui.fillColor,
-      fillOpacity: state.gui.fillOpacity,
-      strokeColor: state.gui.strokeColor,
-      strokeOpacity: state.gui.strokeOpacity,
-      size: s.modulateSize(state.gui.size),
-      shape: state.gui.shape,
-      mirrorX: state.gui.mirrorX,
-      mirrorY: state.gui.mirrorY,
-      isRainbowFill: state.gui.isRainbowFill,
-      isRainbowStroke: state.gui.isRainbowStroke,
-      rainbowFill: rainbowFill.toString(),
-      rainbowStroke: rainbowStroke.toString(),
-    };
-
+    const paintProperties = setupPaintProperties(s, state);
     s.updateDrawing(paintProperties);
     socket.emit("update", paintProperties);
   };
